+++ include/math/sgemm.hpp
@@ -18,8 +18,9 @@
 #include <memory>
 
 namespace tmath {
-void sgemm(const float *A, const float *B, float *C, const int M, const int N, const int K,
+void sgemm(const float *A, const float *B, float *C, const size_t M, const size_t N, const size_t K,
            const bool trans_A = false, const bool trans_B = false);
-void dgemm(const double *A, const double *B, double *C, const int M, const int N, const int K);
+void dgemm(const double *A, const double *B, double *C, const size_t M, const size_t N,
+           const size_t K);
 
 } // namespace tmath
